bugfix: fix compilation on musl libc
isinff is deprecated since C99 and only glibc still implements it
nowadays while musl libc doesn't, replace it with the modern counterpart
and start using the C++ overload.

Meanwhile M_PIf is a glibc 2.35+ macro which doesn't exist on musl libc
math.h, so just use M_PI which seems to work the same in this case.