fix: Correctly handle default configuration when user dirs do not exist
We assumed that the user configuration directory hierarchy always
exists, but with a completely fresh installation, it does not.

Consequently, populating the default configuration silently failed
and the environment is broken.

To resolve this issue, we now check to see if the directory the
configuration file lives in exists, and we create it if it does not.

Additionally, we now support populating a full set of default
configuration in case multiple files and folders are required for
initial configuration.