fix: Drop broken method for seeding default configuration
With b4fa88e9facd312c6944a6e9e44c02bd4ad54ccc, the main method for
initializing default configuration copies over a whole configuration
directory for the purpose of being able to handle a full set of
configuration all at once. This is the much more coherent and
expected method for seeding default configuration.

Thus, we do not need the method for copying a single file anymore,
and this change drops it to simplify the understanding of seeding
configuration in MiracleWM.