cmake: Update CMake version, correctly set up targets, drop useless stuff
Bumping the CMake minimum version ensures we remain compatible for
future deprecations coming.

The new setup for targets ensures the CXX standard is set as part
of the target and also correctly reuses GNUInstallDirs.

The library is also changed to be versioned with the project,
effectively switching to SemVer.

The custom commands produced artifacts that are not even used for
installing the software, so they are dropped.